refactor: 기존 파일을 폴더로 정리하여 구조 개선
